pyproject.toml
[tool.poetry]
name = "souvenir"
version = "1.0.2"
description = "Little CLI program which helps creating and viewing flashcards"
authors = ["Denis Gruzdev <codingjerk@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/codingjerk/souvenir"
repository = "https://github.com/codingjerk/souvenir"
keywords = ["flashcard", "cli", "tool"]
classifiers = [
    "Environment :: Console",
    "Operating System :: OS Independent",
    "Topic :: Utilities",
]

[tool.poetry.scripts]
sv = 'souvenir.__main__:sv'

[tool.poetry.dependencies]
python = "^3.9"
typer = "^0.4.0"
tabulate = "^0.8.9"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
